misc/legacy/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kubelingo"
version = "0.2.0"
description = "CKAD Studying Tool"
readme = "README.md"
authors = [
    { name = "Edward Joseph", email = "aptnative@gmail.com" }
]
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "boto3>=1.26.0",
    "click>=8.0.0",
    "colorama>=0.4.4",
    "kubernetes>=25.0.0",
    "PyYAML>=6.0",
    "questionary>=1.10.0",
    "python-dotenv>=0.21.0",
]

[project.urls]
"Homepage" = "https://github.com/josephedward/kubelingo"

[project.optional-dependencies]
test = [
    "pytest>=6.0",
    "requests>=2.28.0",
]
llm = [
    "llm>=0.14.0",
    "llm-openai>=0.5.0",
    "llm-gemini>=0.5.0",
]

[project.scripts]
kubelingo = "kubelingo.kubelingo:cli"

[tool.pytest.ini_options]
markers = [
    "network: marks tests that require a network connection",
    "e2e: marks end-to-end tests that require a full installation",
]
addopts = "-m 'not network'"
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]